Reworked cast button handling
