+++ app/src/free/java/de/danoeh/antennapod/activity/CastEnabledActivity.java
@@ -2,6 +2,8 @@ package de.danoeh.antennapod.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.view.Menu;
+
 /**
  * Activity that allows for showing the MediaRouter button whenever there's a cast device in the
  * network.
@@ -9,7 +11,7 @@ import androidx.appcompat.app.AppCompatActivity;
 public abstract class CastEnabledActivity extends AppCompatActivity {
     public static final String TAG = "CastEnabledActivity";
 
-    public final void requestCastButton(int showAsAction) {
+    public final void requestCastButton(Menu menu) {
         // no-op
     }
 }
