+++ app/src/main/java/de/danoeh/antennapod/activity/MediaplayerActivity.java
@@ -295,9 +295,7 @@ public abstract class MediaplayerActivity extends CastEnabledActivity implements
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         super.onCreateOptionsMenu(menu);
-        if (Flavors.FLAVOR == Flavors.PLAY) {
-            requestCastButton(MenuItem.SHOW_AS_ACTION_ALWAYS);
-        }
+        requestCastButton(menu);
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.mediaplayer, menu);
         return true;
